+++ src/Core/Attributes/Setting/Color.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+namespace Nadybot\Core\Attributes\Setting;
+
+use Attribute;
+use Nadybot\Core\Attributes\DefineSetting;
+
+#[Attribute(Attribute::TARGET_CLASS|Attribute::IS_REPEATABLE)]
+class Color extends DefineSetting {
+	/**
+	 * @inheritDoc
+	 */
+	public function __construct(
+		public string $name,
+		public string $description,
+		public int|float|string|bool $defaultValue,
+		public string $type='color',
+		public string $mode='edit',
+		public array $options=[],
+		public string $accessLevel='mod',
+		public ?string $help=null,
+	) {
+		$this->type = 'color';
+		if (preg_match("/^#?([0-9a-f]{6})$/i", (string)$defaultValue, $matches)) {
+			$this->defaultValue = "<font color='#{$matches[1]}'>";
+		}
+	}
+}