+++ src/Core/Modules/ALTS/AltsController.php
@@ -45,6 +45,36 @@ use Nadybot\Core\{
 		accessLevel: "all",
 		description: "Declines being the alt of someone else",
 	),
+
+	NCA\Setting\Boolean(
+		name: 'alts_require_confirmation',
+		description: 'Adding alt requires confirmation from alt',
+		defaultValue: true,
+	),
+	NCA\Setting\Boolean(
+		name: 'alts_show_org',
+		description: 'Show the org in the altlist',
+		defaultValue: true,
+	),
+	NCA\Setting\Options(
+		name: 'alts_profession_display',
+		description: 'How to show profession in alts list',
+		defaultValue: 1,
+		options: [
+			'off' => 0,
+			'icon' => 1,
+			'short' => 2,
+			'full' => 3,
+			'icon+short' => 4,
+			'icon+full' => 5,
+		],
+	),
+	NCA\Setting\Options(
+		name: 'alts_sort',
+		description: 'By what to sort the alts list',
+		defaultValue: 'level',
+		options: ["level", "name"],
+	),
 	NCA\ProvidesEvent("alt(add)"),
 	NCA\ProvidesEvent("alt(del)"),
 	NCA\ProvidesEvent("alt(validate)"),
@@ -85,47 +115,6 @@ class AltsController extends ModuleInstance {
 
 	#[NCA\Setup]
 	public function setup(): void {
-		$this->settingManager->add(
-			module: $this->moduleName,
-			name: 'alts_require_confirmation',
-			description: 'Adding alt requires confirmation from alt',
-			mode: 'edit',
-			type: 'bool',
-			value: '1',
-		);
-		$this->settingManager->add(
-			module: $this->moduleName,
-			name: 'alts_show_org',
-			description: 'Show the org in the altlist',
-			mode: 'edit',
-			type: 'bool',
-			value: '1',
-		);
-		$this->settingManager->add(
-			module: $this->moduleName,
-			name: 'alts_profession_display',
-			description: 'How to show profession in alts list',
-			mode: 'edit',
-			type: 'options',
-			value: '1',
-			options: [
-				'off' => 0,
-				'icon' => 1,
-				'short' => 2,
-				'full' => 3,
-				'icon+short' => 4,
-				'icon+full' => 5,
-			],
-		);
-		$this->settingManager->add(
-			module: $this->moduleName,
-			name: 'alts_sort',
-			description: 'By what to sort the alts list',
-			mode: 'edit',
-			type: 'options',
-			value: 'level',
-			options: ["level", "name"],
-		);
 		$this->cacheAlts();
 	}