+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Allow adding settings via Attributes
+
+### Changed
+
+- Database table versions are displayed with a date and time (if possible)
+- Changed the Core module settings to using attributes
+
 ## [6.0.0-beta.1] - 2022-03-22
 
 ### Added